feat(CI): fail CI if one of integration matrix fail (#1703)
* add post status check

* use vanilla status update

* add readme and if always for status job

* decrease the test timeout

* update readme

* create misc folder for tests

* other components folder

* use _ instead space

* ignore .md changes

* skip test for md files
