+++ .github/workflows/test-integration.yml
@@ -1,7 +1,8 @@
-name: Test
-
+name: Integration
 on:
   pull_request:
+    paths-ignore:
+      - '**.md'
   push:
     paths-ignore:
       - '**.md'
@@ -18,12 +19,14 @@ jobs:
       - uses: actions/checkout@v2
       - name: Finding files and store to output
         id: set-matrix
-        run: echo "::set-output name=matrix::$(find ./integration -type d -maxdepth 1 ! -name 'integration' -print | jq -R . | jq -cs .)"
+        run: echo "::set-output name=matrix::$({ cd integration && find . -type d -maxdepth 1 -print; } | tail -n +2 | cut -c 3- | jq -R . | jq -cs .)"
 
   integration:
+    name: test ${{ matrix.test-path }}
     runs-on: ubuntu-latest
     needs: pre-test
     if: fromJSON(needs.pre-test.outputs.matrix)[0] != null
+    continue-on-error: false
     strategy:
       fail-fast: true
       matrix:
@@ -44,4 +47,17 @@ jobs:
       - name: Install Starport
         run: go install ./...
       - name: Run Integration Tests
-        run: go test -v -timeout 120m ${{ matrix.test-path }}
+        run: go test -v -timeout 30m ./integration/${{ matrix.test-path }}
+
+  status:
+    runs-on: ubuntu-latest
+    needs: integration
+    if: always()
+    steps:
+      - name: Update result status
+        run: |
+          if [ "${{ needs.integration.result }}" = "failure" ]; then
+              exit 1
+          else
+              exit 0
+          fi
