+++ .github/workflows/test-lint.yml
@@ -1,7 +1,11 @@
 name: Test
 on:
   pull_request:
+    paths-ignore:
+      - '**.md'
   push:
+    paths-ignore:
+      - '**.md'
     branches:
       - master
       - develop
