+++ .github/workflows/test.yml
@@ -2,7 +2,11 @@ name: Test
 
 on:
   pull_request:
+    paths-ignore:
+      - '**.md'
   push:
+    paths-ignore:
+      - '**.md'
     branches:
       - master
       - develop
