+++ integration/app/cmd_ibc_test.go
@@ -1,7 +1,7 @@
 //go:build !relayer
 // +build !relayer
 
-package integration_test
+package app_test
 
 import (
 	"path/filepath"
