+++ integration/list/cmd_list_test.go
@@ -1,7 +1,7 @@
 //go:build !relayer
 // +build !relayer
 
-package integration_test
+package list_test
 
 import (
 	"path/filepath"
