+++ integration/map/cmd_map_test.go
@@ -1,7 +1,7 @@
 //go:build !relayer
 // +build !relayer
 
-package integration_test
+package map_test
 
 import (
 	"path/filepath"
