+++ integration/other_components/cmd_message_test.go
@@ -1,7 +1,7 @@
 //go:build !relayer
 // +build !relayer
 
-package integration_test
+package other_components_test
 
 import (
 	"path/filepath"
@@ -11,19 +11,19 @@ import (
 	"github.com/tendermint/starport/starport/pkg/cmdrunner/step"
 )
 
-func TestGenerateAnAppWithQuery(t *testing.T) {
+func TestGenerateAnAppWithMessage(t *testing.T) {
 	var (
 		env  = envtest.New(t)
 		path = env.Scaffold("blog")
 	)
 
-	env.Must(env.Exec("create a query",
+	env.Must(env.Exec("create a message",
 		step.NewSteps(step.New(
 			step.Exec(
 				"starport",
 				"s",
-				"query",
-				"foo",
+				"message",
+				"do-foo",
 				"text",
 				"vote:int",
 				"like:bool",
@@ -34,39 +34,36 @@ func TestGenerateAnAppWithQuery(t *testing.T) {
 		)),
 	))
 
-	env.Must(env.Exec("create a query with custom path",
+	env.Must(env.Exec("create a message with custom path",
 		step.NewSteps(step.New(
 			step.Exec(
 				"starport",
 				"s",
-				"query",
-				"AppPath",
+				"message",
+				"app-path",
 				"text",
 				"vote:int",
 				"like:bool",
 				"-r",
 				"foo,bar:int,foobar:bool",
 				"--path",
-				"./blog",
+				"blog",
 			),
 			step.Workdir(filepath.Dir(path)),
 		)),
 	))
 
-	env.Must(env.Exec("create a paginated query",
+	env.Must(env.Exec("should prevent creating an existing message",
 		step.NewSteps(step.New(
-			step.Exec(
-				"starport",
-				"s",
-				"query",
-				"bar",
-				"text",
-				"vote:int",
-				"like:bool",
-				"-r",
-				"foo,bar:int,foobar:bool",
-				"--paginated",
-			),
+			step.Exec("starport", "s", "message", "do-foo", "bar"),
+			step.Workdir(path),
+		)),
+		envtest.ExecShouldError(),
+	))
+
+	env.Must(env.Exec("create a message with a custom signer name",
+		step.NewSteps(step.New(
+			step.Exec("starport", "s", "message", "do-bar", "bar", "--signer", "bar-doer"),
 			step.Workdir(path),
 		)),
 	))
@@ -94,26 +91,11 @@ func TestGenerateAnAppWithQuery(t *testing.T) {
 		)),
 	))
 
-	env.Must(env.Exec("create a query with the custom field type",
+	env.Must(env.Exec("create a message with the custom field type",
 		step.NewSteps(step.New(
-			step.Exec("starport", "s", "query", "foobaz", "bar:CustomType"),
+			step.Exec("starport", "s", "message", "foo-baz", "customField:CustomType"),
 			step.Workdir(path),
 		)),
-	))
-
-	env.Must(env.Exec("create an empty query",
-		step.NewSteps(step.New(
-			step.Exec("starport", "s", "query", "foobar"),
-			step.Workdir(path),
-		)),
-	))
-
-	env.Must(env.Exec("should prevent creating an existing query",
-		step.NewSteps(step.New(
-			step.Exec("starport", "s", "query", "foo", "bar"),
-			step.Workdir(path),
-		)),
-		envtest.ExecShouldError(),
 	))
 
 	env.Must(env.Exec("create a module",
@@ -123,14 +105,15 @@ func TestGenerateAnAppWithQuery(t *testing.T) {
 		)),
 	))
 
-	env.Must(env.Exec("create a query in a module",
+	env.Must(env.Exec("create a message in a module",
 		step.NewSteps(step.New(
 			step.Exec(
 				"starport",
 				"s",
-				"query",
-				"foo",
+				"message",
+				"do-foo",
 				"text",
+				"userIds:array.uint",
 				"--module",
 				"foo",
 				"--desc",
