+++ integration/other_components/cmd_query_test.go
@@ -1,7 +1,7 @@
 //go:build !relayer
 // +build !relayer
 
-package integration_test
+package other_components_test
 
 import (
 	"path/filepath"
@@ -11,19 +11,19 @@ import (
 	"github.com/tendermint/starport/starport/pkg/cmdrunner/step"
 )
 
-func TestGenerateAnAppWithMessage(t *testing.T) {
+func TestGenerateAnAppWithQuery(t *testing.T) {
 	var (
 		env  = envtest.New(t)
 		path = env.Scaffold("blog")
 	)
 
-	env.Must(env.Exec("create a message",
+	env.Must(env.Exec("create a query",
 		step.NewSteps(step.New(
 			step.Exec(
 				"starport",
 				"s",
-				"message",
-				"do-foo",
+				"query",
+				"foo",
 				"text",
 				"vote:int",
 				"like:bool",
@@ -34,36 +34,39 @@ func TestGenerateAnAppWithMessage(t *testing.T) {
 		)),
 	))
 
-	env.Must(env.Exec("create a message with custom path",
+	env.Must(env.Exec("create a query with custom path",
 		step.NewSteps(step.New(
 			step.Exec(
 				"starport",
 				"s",
-				"message",
-				"app-path",
+				"query",
+				"AppPath",
 				"text",
 				"vote:int",
 				"like:bool",
 				"-r",
 				"foo,bar:int,foobar:bool",
 				"--path",
-				"blog",
+				"./blog",
 			),
 			step.Workdir(filepath.Dir(path)),
 		)),
 	))
 
-	env.Must(env.Exec("should prevent creating an existing message",
+	env.Must(env.Exec("create a paginated query",
 		step.NewSteps(step.New(
-			step.Exec("starport", "s", "message", "do-foo", "bar"),
-			step.Workdir(path),
-		)),
-		envtest.ExecShouldError(),
-	))
-
-	env.Must(env.Exec("create a message with a custom signer name",
-		step.NewSteps(step.New(
-			step.Exec("starport", "s", "message", "do-bar", "bar", "--signer", "bar-doer"),
+			step.Exec(
+				"starport",
+				"s",
+				"query",
+				"bar",
+				"text",
+				"vote:int",
+				"like:bool",
+				"-r",
+				"foo,bar:int,foobar:bool",
+				"--paginated",
+			),
 			step.Workdir(path),
 		)),
 	))
@@ -91,11 +94,26 @@ func TestGenerateAnAppWithMessage(t *testing.T) {
 		)),
 	))
 
-	env.Must(env.Exec("create a message with the custom field type",
+	env.Must(env.Exec("create a query with the custom field type",
 		step.NewSteps(step.New(
-			step.Exec("starport", "s", "message", "foo-baz", "customField:CustomType"),
+			step.Exec("starport", "s", "query", "foobaz", "bar:CustomType"),
 			step.Workdir(path),
 		)),
+	))
+
+	env.Must(env.Exec("create an empty query",
+		step.NewSteps(step.New(
+			step.Exec("starport", "s", "query", "foobar"),
+			step.Workdir(path),
+		)),
+	))
+
+	env.Must(env.Exec("should prevent creating an existing query",
+		step.NewSteps(step.New(
+			step.Exec("starport", "s", "query", "foo", "bar"),
+			step.Workdir(path),
+		)),
+		envtest.ExecShouldError(),
 	))
 
 	env.Must(env.Exec("create a module",
@@ -105,15 +123,14 @@ func TestGenerateAnAppWithMessage(t *testing.T) {
 		)),
 	))
 
-	env.Must(env.Exec("create a message in a module",
+	env.Must(env.Exec("create a query in a module",
 		step.NewSteps(step.New(
 			step.Exec(
 				"starport",
 				"s",
-				"message",
-				"do-foo",
+				"query",
+				"foo",
 				"text",
-				"userIds:array.uint",
 				"--module",
 				"foo",
 				"--desc",
